requirements.txt
notion-client
python-dotenv
requests
python-telegram-bot[job-queue]
FinMind
httpx
pandas
tqdm
yfinance
